wip: new invest page
