+++ src/invest/invest-detail/index.ts
@@ -1,0 +1,1 @@
+export * from './invest-detail'
