+++ src/invest/invest-detail/invest-detail.css.ts
@@ -1,0 +1,7 @@
+import { style } from '@vanilla-extract/css'
+
+export const root = style({
+  height: '100%',
+  display: 'flex',
+  flexDirection: 'column',
+})
